hax: Support guest RAM sizes of 4GB or more
Since HAX_VM_IOCTL_ALLOC_RAM takes a 32-bit size, it cannot handle
RAM blocks of 4GB or larger, which is why HAXM can only run guests
with less than 4GB of RAM. Solve this problem by utilizing the new
HAXM API, HAX_VM_IOCTL_ADD_RAMBLOCK, which takes a 64-bit size, to
register RAM blocks with the HAXM kernel module. The new API is
first added in HAXM 7.0.0, and its availablility can be confirmed
by the presence of the HAX_CAP_64BIT_RAMBLOCK capability flag.

With this patch, guests with >= 7GB of RAM still won't boot under
HAXM, due to a similar size limitation in HAX_VM_IOCTL_SET_RAM.
That will be addressed by another patch.

Bug: https://bugs.launchpad.net/qemu/+bug/1735576

+++ include/sysemu/hax.h
@@ -32,7 +32,7 @@ int hax_smp_cpu_exec(CPUState *cpu);
 int hax_vcpu_emulation_mode(CPUState *cpu);
 int hax_stop_emulation(CPUState *cpu);
 int hax_stop_translate(CPUState *cpu);
-int hax_populate_ram(uint64_t va, uint32_t size);
+int hax_populate_ram(uint64_t va, uint64_t size);
 
 void hax_cpu_synchronize_state(CPUState *cpu);
 void hax_cpu_synchronize_post_reset(CPUState *cpu);
